+++ yodl/src/file/filecopy2offset.c
@@ -16,7 +16,7 @@ void file_copy2offset(FILE *dest, FILE *src, long src_end)
             nread = fread(buffer, 1, BLOCK_FILE, src);
             if (nread <= 0)
                 return;
-            fwrite(buffer, (unsigned)nread, 1, dest);
+            fwrite(buffer, (size_t)nread, 1, dest);
         }
     }
 
@@ -35,10 +35,10 @@ void file_copy2offset(FILE *dest, FILE *src, long src_end)
         tocopy <= BLOCK_FILE                    /* at most 1 block waits    */
         &&                                      /* AND */
                                                 /* we read the req'd chars  */
-        fread(buffer, (unsigned)tocopy, 1, src) == 1     
+        fread(buffer, (size_t)tocopy, 1, src) == 1     
         &&                                      /* AND */
                                                 /* we wrote the req'd chars */
-        fwrite(buffer, (unsigned)tocopy, 1, dest) == 1   
+        fwrite(buffer, (size_t)tocopy, 1, dest) == 1   
     )
     {
         free(buffer);
