+++ yodl/src/hashmap/hashmaperase.c
@@ -2,7 +2,7 @@
 
 Result hashmap_erase(register HashMap *map, char const *name, SymbolType type)
 {
-    register unsigned idx = hm_find_idx(map, name, type);
+    register size_t idx = hm_find_idx(map, name, type);
 
     if (idx == UFAILED)
         return idx;
