+++ yodl/src/hashmap/hashmapfind.c
@@ -4,7 +4,7 @@
 
 HashItem *hashmap_find(register HashMap *map, char const *key, SymbolType type)
 {
-    unsigned idx = hm_find_idx(map, key, type);
+    size_t idx = hm_find_idx(map, key, type);
 
     return idx == UFAILED ? PFAILED : map->d_map[idx];
 }
