+++ yodl/src/lines/lines.h
@@ -8,24 +8,24 @@
 typedef struct
 {
     char **d_str;
-    unsigned d_size;
-    unsigned d_capacity;
+    size_t d_size;
+    size_t d_capacity;
 }
 Lines;
 
 void        lines_add(Lines *lines, char const *line);  /* copies line      */
-char const *lines_at(Lines *lines, unsigned idx);   /* 0: not existing      */
+char const *lines_at(Lines *lines, size_t idx);   /* 0: not existing      */
                                                     /* Not yet implemented  */
-unsigned    lines_contains(Lines *lp, char const *target);
+size_t      lines_contains(Lines *lp, char const *target);
 void        lines_construct(Lines *lines);
 void        lines_destroy(Lines *lines);
                                                     /* FAILED or index      */
-unsigned    lines_find(char const *target, char const **arr,
-                       unsigned arrsize);
+size_t      lines_find(char const *target, char const **arr,
+                       size_t arrsize);
                                                     /* adds sprint-ed       */
 void        lines_format(Lines *lines, char const *line, ...)
             ATTRIBUTE_FORMAT_PRINTF(2, 3);
 void        lines_grab(Lines *lines, char *line);       /* grabs line       */
-unsigned    lines_size(Lines *lines);                   /* n lines stored   */
+size_t      lines_size(Lines *lines);                   /* n lines stored   */
 
 #endif
