+++ yodl/src/lines/linessize.c
@@ -1,6 +1,6 @@
 #include "lines.h"
 
-unsigned lines_size(register Lines *lines)
+size_t lines_size(register Lines *lines)
 {
     return lines->d_size;
 }
