+++ yodl/src/message/message.h
@@ -55,15 +55,15 @@ typedef enum
 }
 MESSAGE_SEVERITY;
 
-#define UMSG_ALL ((unsigned)MSG_ALL)
+#define UMSG_ALL ((size_t  )MSG_ALL)
 
 typedef struct
 {
-    unsigned            d_max_errors;
-    unsigned            d_errors;
+    size_t              d_max_errors;
+    size_t              d_errors;
     char                *d_program_name;
     char                *d_filename;
-    unsigned            d_lineno;
+    size_t              d_lineno;
     MESSAGE_SEVERITY    d_severity;
     MESSAGE_SEVERITY    d_last_show;
     bool                d_warn;
@@ -78,9 +78,9 @@ void    message_error(char const *fmt,...) ATTRIBUTE_FORMAT_PRINTF(1, 2);
 void    message_optarg(char const *opt);
 bool    message_show(MESSAGE_SEVERITY level);
 void    message_setfilename(char const *newname);
-void    message_setlineno(unsigned lineno);
-void    message_setmaxerrors(unsigned max);
-void    message_setmask(unsigned mask);
+void    message_setlineno(size_t   lineno);
+void    message_setmaxerrors(size_t   max);
+void    message_setmask(size_t   mask);
 void    message_setseverity(MESSAGE_SEVERITY level);
 void    message_setverbosity(int mode, char *arg);  /* arg may be modified  */
 void    message_setwarn(bool trueIsOn);
@@ -91,7 +91,7 @@ void    warning(char const* format, ...) ATTRIBUTE_FORMAT_PRINTF(1, 2);
 
 bool                message_errors(void);
 char const         *message_filename(void);
-unsigned            message_lineno(void);
+size_t              message_lineno(void);
 char const         *message_programname(void);
 char const         *message_verbosity(void);       /* returns (hex) mask   */
 char const         *message_version(void);
