+++ yodl/src/message/messagesetlineno.c
@@ -1,6 +1,6 @@
 #include "message.ih"
 
-void message_setlineno(unsigned line)
+void message_setlineno(size_t line)
 {
     m_message.d_lineno = line;
 }
