+++ yodl/src/message/messagesetmaxerrors.c
@@ -1,6 +1,6 @@
 #include "message.ih"
 
-void message_setmaxerrors(unsigned max)
+void message_setmaxerrors(size_t max)
 {
     m_message.d_max_errors = max;
 }
