+++ yodl/src/new/newcalloc.c
@@ -1,6 +1,6 @@
 #include "new.ih"
 
-void *new_calloc(unsigned nElements, unsigned sizeofElement)
+void *new_calloc(size_t nElements, size_t sizeofElement)
 {
     nElements *= sizeofElement;
     return memset(n_malloc(nElements), 0, nElements);
